refactor: merge ui components
