+++ packages/component/src/stories/page-list.stories.tsx
@@ -10,7 +10,7 @@ import type {
 } from '../components/page-list/all-page';
 import { PageListTrashView } from '../components/page-list/all-page';
 import PageList from '../components/page-list/all-page';
-import { NewPageButton } from '../components/page-list/new-page-buttton';
+import { NewPageButton } from '../components/page-list/components/new-page-buttton';
 import type { OperationCellProps } from '../components/page-list/operation-cell';
 import { OperationCell } from '../components/page-list/operation-cell';
 import { toast } from '../ui/toast';
